post-rebase fixes
