+++ next.config.ts
@@ -24,7 +24,6 @@ const nextConfig: NextConfig = {
   },
   experimental: {
     nodeMiddleware: true,
-    ppr: true,
   },
 };
